+++ metplus/wrappers/ascii2nc_wrapper.py
@@ -29,7 +29,7 @@ class ASCII2NCWrapper(CommandBuilder):
         'METPLUS_TIME_SUMMARY_DICT',
     ]
 
-    def __init__(self, config, instance=None, config_overrides={}):
+    def __init__(self, config, instance=None, config_overrides=None):
         self.app_name = "ascii2nc"
         self.app_path = os.path.join(config.getdir('MET_BIN_DIR', ''),
                                      self.app_name)