+++ metplus/wrappers/gen_vx_mask_wrapper.py
@@ -25,7 +25,7 @@ from ..util import do_string_sub
 
 class GenVxMaskWrapper(CommandBuilder):
 
-    def __init__(self, config, instance=None, config_overrides={}):
+    def __init__(self, config, instance=None, config_overrides=None):
         self.app_name = "gen_vx_mask"
         self.app_path = os.path.join(config.getdir('MET_BIN_DIR', ''),
                                      self.app_name)