+++ metplus/wrappers/pb2nc_wrapper.py
@@ -36,7 +36,7 @@ class PB2NCWrapper(CommandBuilder):
         'METPLUS_QUALITY_MARK_THRESH',
     ]
 
-    def __init__(self, config, instance=None, config_overrides={}):
+    def __init__(self, config, instance=None, config_overrides=None):
         self.app_name = 'pb2nc'
         self.app_path = os.path.join(config.getdir('MET_BIN_DIR', ''),
                                      self.app_name)